Refactoring Consolidating shared components Fixing styles Adding Missing links
