+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "license": "Apache-2.0",
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/IBM/watson-discovery-news.git"
+    "type": "git",
+    "url": "https://github.com/IBM/watson-discovery-news.git"
   },
   "scripts": {
     "start": "node index.js",
@@ -19,6 +19,9 @@
   "dependencies": {
     "babelify": "^7.3.0",
     "bluebird": "^3.5.0",
+    "botkit": "^0.5.4",
+    "cf-deployment-tracker-client": "*",
+    "cfenv": "^1.0.3",
     "dotenv": "^4.0.0",
     "express": "^4.15.3",
     "express-browserify": "^1.0.2",
@@ -31,8 +34,6 @@
     "recharts": "^1.0.0-alpha.0",
     "rss": "^1.2.2",
     "uglifyify": "^3.0.4",
-    "cf-deployment-tracker-client": "*",
-    "cfenv": "^1.0.3",
     "vcap_services": "^0.2.0",
     "watson-developer-cloud": "^2.32.1",
     "watson-react-components": "^0.6.9"
