+++ server/search/query-builder.js
@@ -17,7 +17,6 @@ module.exports = {
       count: 10,
       sort: '-_score',
       filter: 'blekko.hostrank>500',
-      aggregation: aggregations,
       return: 'enrichedTitle.text,text,title,url,host,blekko.chrondate,blekko.hostrank,score,id,entities.text,docSentiment.type',
       aggregations
     }, queryOpts);
